refactor: Twitchスコープ定数をクライアント安全なモジュールに分離
ChannelPointSettings (use client) が @/lib/twitch/auth 経由で
@/lib/env-validation を transitively import し、ブラウザで
サーバー専用シークレットの top-level 検証が走ると落ちる可能性があった。

- src/lib/twitch/scopes.ts を新設し AUTH_SCOPES/ADDITIONAL_SCOPES/
  CHANNEL_POINT_SCOPES を移動（env 依存なし、client-safe）
- auth.ts は scopes.ts を内部 import するだけの server-only モジュールに
- client / server の全呼び出し元を scopes.ts に差し替え
- テストの auth モックから scope 定義を削除し、実 scopes.ts を利用
