+++ src/app/api/auth/twitch/check-subscription/route.ts
@@ -4,7 +4,7 @@ import { getSession } from '@/lib/session'
 import { checkRateLimit, rateLimits, getRateLimitIdentifier } from '@/lib/rate-limit'
 import { hasScope, removeScope } from '@/lib/twitch/token-manager'
 import { checkTwitchSubViaApi, isTwitchSubCheckEnabled } from '@/lib/twitch/sub-check'
-import { ADDITIONAL_SCOPES } from '@/lib/twitch/auth'
+import { ADDITIONAL_SCOPES } from '@/lib/twitch/scopes'
 import { ERROR_MESSAGES } from '@/lib/constants'
 import { getSupabaseAdmin } from '@/lib/supabase/admin'
 import { logger } from '@/lib/logger'
